Some UI Elements
Forgot to add the plain black box. These are just some I worked on for fun, do not have to use. Can change the pallet. To quickly put them in if you want to see them just go to the prefab, isolate it, and then change the sprites. You might need to change it from simple to spliced. It may come up with an error saying no boarder but I'll fix that when I get in.